agregando codigo de eleiminar y actualizar receta

+++ api/src/controllers/recipes/deleteRecipe.js
@@ -4,6 +4,12 @@ const { response } = require('../../../utils')
 module.exports = async (req, res) => {
   const { id } = req.params
 
+  const recipeId = await Recipe.findByPk(id)
+
+  if(!recipeId) {
+    return response(res, 404, { msg: 'Recipe not found' });
+  }
+
   await Recipe.destroy({
     where: {
       id: id
